testing

+++ database/dbHelper/user.go
@@ -16,18 +16,14 @@ func CreateUser(user *models.User) error {
 	`
 
 	return migrations.DB.QueryRow(
-		query,
-		user.Name,
-		user.Email,
-		user.Password,
+		query, user.Name, user.Email, user.Password,
 	).Scan(&user.UserID)
 }
 
 func DeleteUser(userID uuid.UUID) (int64, error) {
 
 	result, err := migrations.DB.Exec(
-		`DELETE FROM users WHERE user_id=$1`,
-		userID,
+		`DELETE FROM users WHERE user_id=$1`, userID,
 	)
 
 	if err != nil {
@@ -41,13 +37,8 @@ func GetUserByEmail(email string) (models.User, error) {
 
 	var user models.User
 
-	err := migrations.DB.Get(
-		&user,
-		`
-		SELECT user_id,name,email,password
-		FROM users
-		WHERE email=$1
-		`,
+	err := migrations.DB.Get(&user, `
+		SELECT user_id,name,email,password FROM users WHERE email=$1`,
 		email,
 	)
